refactor: extract instance history search form and reuse tree node in filtered list
Move the inline react-final-form block from ElementInstanceLog into a
dedicated SearchForm component so the form logic is testable in isolation
and the orchestrator only reads the submitted search from the URL param.

Reuse the tree's styled TreeNode and ElementInstanceIcon (with
$hasLeftMargin) in FilteredElementInstancesList so both views are visually
aligned by construction; remove the duplicate SearchResultNode styled
component.

Also fixes a bug where an active search filter was showing the filtered
list instead of the tree in modification mode: gate the body switch on
isFiltered && !isModificationModeEnabled.
