+++ operate/client/src/App/ProcessInstance/ElementInstanceLog/FilteredElementInstancesList/index.tsx
@@ -20,7 +20,7 @@ import {ErrorMessage} from 'modules/components/ErrorMessage';
 import {InfiniteScroller} from 'modules/components/InfiniteScroller';
 import {Skeleton} from '../Skeleton';
 import {Bar} from '../ElementInstancesTree/Bar';
-import {ElementInstanceIcon} from 'modules/components/ElementInstanceIcon';
+import {TreeNode, ElementInstanceIcon} from '../ElementInstancesTree/styled';
 import {useProcessInstanceElementSelection} from 'modules/hooks/useProcessInstanceElementSelection';
 import {isMultiInstance} from 'modules/bpmn-js/utils/isMultiInstance';
 import {tracking} from 'modules/tracking';
@@ -30,12 +30,7 @@ import {useDashboardScrollPagination} from 'App/Dashboard/useDashboardScrollPagi
 import {escapeLikePattern} from 'modules/utils/escapeLikePattern';
 import {isRequestError} from 'modules/request';
 import {HTTP_STATUS_FORBIDDEN} from 'modules/constants/statusCode';
-import {
-  ScrollContainer,
-  SearchResultNode,
-  StatusRegion,
-  EmptyStateContainer,
-} from './styled';
+import {ScrollContainer, StatusRegion, EmptyStateContainer} from './styled';
 
 const PAGE_LIMIT = 50;
 
@@ -58,13 +53,14 @@ const Row: React.FC<RowProps> = observer(({item, businessObjects}) => {
   });
 
   return (
-    <SearchResultNode
+    <TreeNode
       id={item.elementInstanceKey}
       value={item.elementInstanceKey}
       data-testid={`search-result-${item.elementInstanceKey}`}
       aria-label={item.elementName ?? item.elementId}
       renderIcon={() => (
         <ElementInstanceIcon
+          $hasLeftMargin
           diagramBusinessObject={businessObject}
           isRootProcess={item.type === 'PROCESS'}
         />
